# services/price-service/docs/swagger.yaml
basePath: /internal
definitions:
...
  handlers.IngestionChunk:
    properties:
      chunkIndex:
        type: integer
      createdAt:
        type: string
      endRow:
        type: integer
      errorCount:
        type: integer
      fileId:
        type: string
      id:
        type: string
      persistedCount:
        type: integer
      processedAt:
        type: string
      r2Key:
        type: string
      rowCount:
        type: integer
      startRow:
        type: integer
      status:
        type: string
    type: object
  handlers.IngestionError:
    properties:
      chunkId:
        type: string
      createdAt:
        type: string
      entryId:
        type: string
      errorDetails:
        type: string
      errorMessage:
        type: string
      errorType:
        type: string
      fileId:
        type: string
      id:
        type: string
      runId:
        type: string
      severity:
        type: string
    type: object
  handlers.IngestionFile:
    properties:
      chunkSize:
        type: integer
      createdAt:
        type: string
      entryCount:
        type: integer
      failedRows:
        type: integer
      fileHash:
        type: string
      fileSize:
        type: integer
      fileType:
        type: string
      filename:
        type: string
      id:
        type: string
      metadata:
        type: string
      persistedCount:
        type: integer
      priceChanges:
        type: integer
      processedAt:
        type: string
      processedChunks:
        type: integer
      rowCount:
        type: integer
      runId:
        type: string
      status:
        type: string
      statusReason:
        type: string
      statusSeverity:
        type: string
      statusType:
        type: string
      storeCount:
        type: integer
      totalChunks:
        type: integer
      warningRows:
        type: integer
    type: object
...
  handlers.ListChunksResponse:
    properties:
      chunks:
        items:
          $ref: '#/definitions/handlers.IngestionChunk'
        type: array
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  handlers.ListErrorsResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/handlers.IngestionError'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: Internal API for price data management, ingestion monitoring, and basket
    optimization.
  title: Price Service API
  version: "1.0"
paths:
...
  /internal/ingestion/files/{fileId}:
    get:
      consumes:
      - application/json
      description: Returns a single ingestion file by its ID
      parameters:
      - description: File ID
        in: path
        name: fileId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.IngestionFile'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: File not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get ingestion file
      tags:
      - ingestion
  /internal/ingestion/files/{fileId}/chunks:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of chunks for a specific ingestion file
      parameters:
      - description: File ID
        in: path
        name: fileId
        required: true
        type: string
      - description: Chunk status
        enum:
        - pending
        - processing
        - completed
        - failed
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListChunksResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List ingestion chunks
      tags:
      - ingestion
  /internal/ingestion/files/{fileId}/errors:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of errors for a specific ingestion file
      parameters:
      - description: File ID
        in: path
        name: fileId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListErrorsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List ingestion errors for file
      tags:
      - ingestion
...
swagger: "2.0"